Accessible labels for form comboboxes
The label for attribute values for the border color and background
color edit fields (Table Properties - Advanced tab) did not match the
corresponding id attribute values of the input fields.

This change ensures the for attribute value matches the id attribute
of the INPUT field, but adding explicit support for the trick used by
comboboxes for pointing to their INPUT element.